backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
langchain==0.3.27

langgraph==0.6.7
pydantic==2.11.7
python-dotenv==1.0.0
langchain-google-genai == 2.1.10
requests==2.31.0